Migrate code from ChaseAgent to OpencogAgent
